# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/branch/add": {
            "post": {
                "description": "This API Creates a new branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branch"
                ],
                "summary": "This API Creates a new branch",
                "parameters": [
                    {
                        "description": "New Branch",
                        "name": "premium",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FPBranch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPBranch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/branch/all/{spId}": {
            "get": {
                "description": "This API List all Branches for a Service Provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branch"
                ],
                "summary": "This API List all Branches for a Service Provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of Service Provider",
                        "name": "spId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FPBranch"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/branch/delete/{branchId}/{spId}": {
            "delete": {
                "description": "This API Delete branch by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branch"
                ],
                "summary": "This API Delete branch by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of Branch",
                        "name": "branchId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of ServiceProvider",
                        "name": "spId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/branch/find/{branchId}/{spId}": {
            "get": {
                "description": "This API Finds branch by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branch"
                ],
                "summary": "This API Finds branch by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of branch",
                        "name": "branchId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of ServiceProvider",
                        "name": "spId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPBranch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/branch/update": {
            "put": {
                "description": "This API Update existing branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branch"
                ],
                "summary": "This API Update existing branch",
                "parameters": [
                    {
                        "description": "Update Branch",
                        "name": "branch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FPBranch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPBranch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.FPAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.FPBranch": {
            "type": "object",
            "properties": {
                "fpbranchID": {
                    "type": "string"
                },
                "fpbranchName": {
                    "type": "string"
                },
                "fpbranchServiceProviderId": {
                    "type": "string"
                }
            }
        },
...
    }
}